Add missing functions to parking_lot API wrappers
